+++ packages/orq-rc/src/lib/config.ts
@@ -58,7 +58,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.12.0-rc.40",
+  sdkVersion: "4.12.0-rc.41",
   genVersion: "2.918.3",
-  userAgent: "speakeasy-sdk/typescript 4.12.0-rc.40 2.918.3 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 4.12.0-rc.41 2.918.3 2.0 @orq-ai/node",
 } as const;