+++ packages/orq-rc/RELEASES.md
@@ -10939,3 +10939,13 @@ Based on:
 - [typescript v4.12.0-rc.40] packages/orq-rc
 ### Releases
 - [NPM v4.12.0-rc.40] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.40 - packages/orq-rc
+
+## 2026-07-20 13:04:41
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.2 (2.918.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.12.0-rc.41] packages/orq-rc
+### Releases
+- [NPM v4.12.0-rc.41] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.41 - packages/orq-rc